Moved work time DTOs from the place resource to the work time resource.
