+++ src/place/place.service.ts
@@ -8,14 +8,14 @@ import { Place } from './entities/place.entity';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { generateBadRequestException } from 'src/common/generate-exception';
-import { CreatePlaceDto } from './dto/place/create-place.dto';
+import { CreatePlaceDto } from './dto/create-place.dto';
 import { AddressService } from 'src/address/address.service';
 import { User } from 'src/user/entities/user.entity';
 import { PlaceType } from './types/place';
-import { CreateWorkTimeDto } from './dto/work-time/create-work-time.dto';
-import { UpdatePlaceDto } from './dto/place/update-place.dto';
+import { UpdatePlaceDto } from './dto/update-place.dto';
 import { WorkTimeService } from 'src/work-time/work-time.service';
 import { UserService } from 'src/user/user.service';
+import { CreateWorkTimeDto } from 'src/work-time/dto/create-work-time.dto';
 
 @Injectable()
 export class PlaceService {
