+++ src/user/dto/update-user.dto.ts
@@ -1,7 +1,7 @@
 import { OmitType, PartialType } from '@nestjs/mapped-types';
 import { CreateUserDto } from './create-user.dto';
 import { IsNotEmptyObject, IsOptional } from 'class-validator';
-import { UpdateWorkTimeDto } from 'src/place/dto/work-time/update-work-time.dto';
+import { UpdateWorkTimeDto } from 'src/work-time/dto/update-work-time.dto';
 
 export class UpdateUserDto extends PartialType(
   OmitType(CreateUserDto, ['password', 'workTime']),
